requirements.txt
streamlit
PyPDF2
python-docx
openai
python-dotenv
pandas
bcrypt
PyYAML
